refactor: Update certificate.php to display default signee name if not provided
The code changes in certificate.php modify the signee name display logic. If the signee name is not provided in the $academy_head array, the default name "Tolulope Adeyemo" will be displayed. This ensures that a signee name is always shown in the certificate, improving the user experience in the Academy Africa theme.