refactor: Update padding and CSS for contact-us section
The CSS changes in main.css, contact-us.css, and contact-us.scss update the padding values for the .contact-us class from 60px to 24px. This modification reduces the padding and improves the spacing of the contact-us section in the Academy Africa theme. Additionally, the CSS changes in main.css and _footer.scss add the "white-space: nowrap" property to the links, preventing them from wrapping to the next line. These modifications enhance the readability and user experience of the contact-us section and links in the theme.